move global functions to util folder and fix imports
